requirements.txt
Django==2.1.3
djangorestframework==3.9.0

pylint==2.1.1
psycopg2-binary ==2.7.6.1
Pillow == 5.3.0
channels==1.1.8
daphne==1.4.2
ipython==7.1.1

asgi-redis==1.4.3
django-rest-swagger==2.2.0
django-filter==2.0.0

celery==4.1.0
flower==0.9.2
redis==2.10.6
kombu==4.1.0

django-ckeditor==5.6.1
django-js-asset==1.1.0

django-tags-input==4.4.2
raven==6.7.0

uwsgi==2.0.17.1

django-webtest==1.9.4
coverage==4.5.2
django-discover-runner==1.0
factory-boy==2.11.1
